backend/requirements.txt
alembic==1.10.3
bcrypt==4.0.1
cffi==1.15.1
Click==8.1.3
Flask==2.2.3
Flask-Bcrypt==1.0.1
Flask-Cors==3.0.10
Flask-JWT-Extended==4.4.4
Flask-Migrate==2.7.0
Flask-Script==2.0.6
flask_script_extras==0.1.4
Flask-SQLAlchemy==3.0.3
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.4
MarkupSafe==2.1.2
marshmallow==3.19.0
pycparser==2.21
pycryptodome==3.17
PyJWT==2.6.0
python-dateutil==2.8.2
python-editor==1.0.4
six==1.16.0
SQLAlchemy==2.0.9
SQLAlchemy-Utils==0.41.0
Werkzeug==2.2.3
